fix(agents-list): the local leg read the CONTAINER's HOME, so a container saw 1 agent of 122 (#1193)
_agent_list_discover hardcoded Path.home() / ".scitex" / "agent-container" /
"agents". That is right on a bare host and wrong inside a container, where
Path.home() is the agent's own home (/home/agent) and holds no agents tree at
all. The walk skipped its only real root, so the loop that skips a
non-directory root skipped the entire search.

MEASURED on scitex-compute-04, same container, before vs after:

    BEFORE (hardcoded Path.home()):    1 agent discovered
    AFTER  (resolver-based):         122 agents discovered
    is the querying agent in its own listing?   False -> True

The negative control -- restoring the hardcoded root -- reproduces the old
result exactly, so the green is not vacuous.

WHY IT IS A DESIGN BUG, NOT A TYPO. The runtime already exports
SCITEX_AGENT_CONTAINER_YAML_DIRS pointing at the operator's bind-mounted tree
for precisely this reason, and config._resolve._search_dirs already honours it
-- which is why "sac agents find" and "sac agents start" could see agents that
"sac agents list" swore did not exist. Two commands in one CLI, one truth, one
of them wrong. The new _agent_list_roots.user_scope_roots() routes the walk
through that same resolver.

WHY IT MATTERED. The failure produces a CONFIDENT EMPTY ANSWER, which reads
exactly like a finding, and nobody investigates a zero. It has now caught five
agents: three on 2026-08-09 concluded the fleet registry had been wiped and two
escalated it as P1 data loss against a healthy database; one reported "107 of
123 agents are down" to the operator (since withdrawn); one was a single
self-check away from telling him three peers had abandoned their work. The
listing also reported the agent MAKING THE QUERY as not running.

ALSO IN THIS CHANGE, forced by the 512-line cap and worth having on its own
merits: ~250 lines of REMOTE network probing lived in a module named _discover
(a local filesystem walk). Extracted to _agent_list_remote_rows.py.
_agent_list_discover re-exports remote_instance_rows AND the two
underscore-prefixed probes, because the suite imports those from the old path
-- they are part of the de-facto contract regardless of the leading underscore.

The resolver is an injected parameter rather than something a test rewrites at
runtime, so the fallback branch is reached the same way in a test as in
production.

Tests: 279 passed in tests/scitex_agent_container/cli_pkg/_helpers/, including
two new ones pinning the container shape and the resolver-failure fallback.
